+++ src/main/java/org/folio/rest/converter/titles/TitleCollectionResultConverter.java
@@ -19,6 +19,7 @@ public class TitleCollectionResultConverter implements Converter<Titles, TitleCo
 
     return TitleCollectionResult.builder()
       .titleResults(titleResults)
+      .facets(titles.getFacets())
       .totalResults(titles.getTotalResults()).build();
   }